Tests/LibWeb/Text/expected/input-appearance-none-normalized-table-display-height.txt

Normalized table-cell height matches block height: true
